# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "MSGNS API v2",
        "description": "API v2 for the MSGNS application. Provides endpoints for Identity management (authentication, user profile, administration) and Products (product types).",
        "version": "2.0.0"
    },
    "servers": [
        {
            "url": "/api/v2",
            "description": "API v2 base URL"
        }
    ],
    "paths": {
        "/api/v2/ai/responses": {
            "get": {
                "tags": [
                    "AI Responses"
                ],
                "summary": "List AI responses for the authenticated user",
                "description": "Returns a paginated list of AI responses scoped to the authenticated user. Filterable by status and product type.",
                "operationId": "listAiResponses",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "pending",
                                "approved",
                                "edited",
                                "rejected",
                                "applied",
                                "expired"
                            ]
                        }
                    },
                    {
                        "name": "product_type",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "minimum": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "maximum": 100,
                            "minimum": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Paginated list of AI responses",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/AiResponseResource"
                                            }
                                        },
                                        "meta": {
                                            "properties": {
                                                "currentPage": {
                                                    "type": "integer"
                                                },
                                                "lastPage": {
                                                    "type": "integer"
                                                },
                                                "perPage": {
                                                    "type": "integer"
                                                },
                                                "total": {
                                                    "type": "integer"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v2/ai/responses/{id}/approve": {
            "patch": {
                "tags": [
                    "AI Responses"
                ],
                "summary": "Approve an AI response",
                "description": "Transitions an AI response from pending or edited to approved.",
                "operationId": "approveAiResponse",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Response approved"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "AI response not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DomainError"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Invalid status transition",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DomainError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v2/ai/responses/{id}/edit": {
            "patch": {
                "tags": [
                    "AI Responses"
                ],
                "summary": "Edit an AI response",
                "description": "Stores edited content and transitions the response to edited status. Idempotent if already edited.",
                "operationId": "editAiResponse",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "edited_content"
                                ],
                                "properties": {
                                    "edited_content": {
                                        "type": "string",
                                        "maxLength": 10000,
                                        "minLength": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Response edited"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "AI response not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DomainError"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error or invalid status transition",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DomainError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v2/ai/responses/{id}/reject": {
            "patch": {
                "tags": [
                    "AI Responses"
                ],
                "summary": "Reject an AI response",
                "description": "Transitions an AI response from pending or edited to rejected (terminal state).",
                "operationId": "rejectAiResponse",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Response rejected"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "AI response not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DomainError"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Invalid status transition",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DomainError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v2/ai/responses/{id}/apply": {
            "post": {
                "tags": [
                    "AI Responses"
                ],
                "summary": "Apply an AI response to the target platform",
                "description": "Transitions an approved response to applied and delegates publishing to the appropriate applier (Google Reviews, Instagram). Transactional — rolls back if publishing fails.",
                "operationId": "applyAiResponse",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Response applied to platform"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "AI response not found or no applier for product type",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DomainError"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Invalid status transition",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DomainError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "DomainError": {
                "required": [
                    "error"
                ],
                "properties": {
                    "error": {
                        "required": [
                            "code"
                        ],
                        "properties": {
                            "code": {
                                "type": "string"
                            },
                            "context": {
                                "type": "object"
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
...
            "GenerationHistorySummaryItem": {
                "required": [
                    "type_code",
                    "type_name",
                    "quantity"
                ],
                "properties": {
                    "type_code": {
                        "type": "string"
                    },
                    "type_name": {
                        "type": "string"
                    },
                    "quantity": {
                        "type": "integer"
                    },
                    "size": {
                        "type": "string",
                        "nullable": true
                    },
                    "description": {
                        "type": "string",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "GenerationHistoryListItem": {
                "required": [
                    "id",
                    "generated_at",
                    "total_count",
                    "summary"
                ],
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "generated_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "total_count": {
                        "type": "integer"
                    },
                    "summary": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/GenerationHistorySummaryItem"
                        }
                    },
                    "generated_by": {
                        "properties": {
                            "id": {
                                "type": "integer"
                            },
                            "email": {
                                "type": "string",
                                "format": "email"
                            }
                        },
                        "type": "object",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "AiResponseResource": {
                "required": [
                    "id",
                    "product_type",
                    "product_id",
                    "ai_content",
                    "status",
                    "expires_at"
                ],
                "type": "object",
                "allOf": [
                    {
                        "$ref": "#/components/schemas/GenerationHistoryListItem"
                    },
                    {
                        "properties": {
                            "id": {
                                "type": "string",
                                "format": "uuid"
                            },
                            "product_type": {
                                "type": "string"
                            },
                            "product_id": {
                                "type": "integer"
                            },
                            "ai_content": {
                                "type": "string"
                            },
                            "edited_content": {
                                "type": "string",
                                "nullable": true
                            },
                            "status": {
                                "type": "string",
                                "enum": [
                                    "pending",
                                    "approved",
                                    "edited",
                                    "rejected",
                                    "applied",
                                    "expired"
                                ]
                            },
                            "expires_at": {
                                "type": "string",
                                "format": "date-time"
                            },
                            "applied_at": {
                                "type": "string",
                                "format": "date-time",
                                "nullable": true
                            },
                            "created_at": {
                                "type": "string",
                                "format": "date-time"
                            },
                            "updated_at": {
                                "type": "string",
                                "format": "date-time"
                            }
                        },
                        "type": "object"
                    }
                ]
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "Sanctum",
                "description": "Enter your Sanctum access token in the format: Bearer <token>"
            }
        }
    },
...
    "security": [
        {
            "bearerAuth": []
        }
    ]
}